requirements.txt
alembic==1.17.0
asgiref==3.10.0
bibtexparser==1.4.3
black==25.9.0
blinker==1.9.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
dominate==2.9.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
identify==2.6.15
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pyparsing==3.2.3
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.35.3
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
